Create cafData.R
Getting, cleaning and tidying data from the World Bank Org. about Central African Republic (CAR, designed by CAF in WB database)
The goal is to go from a csv file of 1335 rows and 58 columns, written in french, a bit messy with lot of missing values, to a complete data frame, "clean" and exploitable. Ultimately, a motion chart with Google Visualization package gives a dynamic approach on the evolution of the different indicators of the CAF since from independence in 1960 until 2013.